Add makefile + jarjar rule for okhttp.
Created a new path android/ that contains android
specific patches to this code. This is not a long term approach,
simply a stop gap measure to get this codebase compiling and usable
under android. For now, this path contains changes to Libcore.java
to remove OpenJDK dependencies.

This patchset also updates the snapshot to upstream revision
2a05f0e18f0ae67.

Change-Id: Ic1ef8245109e07c7cf9f59c1e7881b86d357d74f