Fix usage DeflaterOutputStream constructor on android
This API is hidden and unbundled apps that attempt to use
it on android will not compile. Also added a
LOCAL_SDK_VERSION to the make file to ensure that it can
compile against an SDK. All uses of hidden / private APIs are now
in Libcore.java so this should make it easier for us to move
forward with building an interface for okhttp to use these
APIs.

Change-Id: Ie7fc097b84445d55745a94ce048dd46590887d80