external/okhttp changes to support connection pool flushing
When the preferred network / default network for a device
changes we must flush any pooled connections.

This change has a known issue which will be fixed in a
later commit: Any connections that are in use when
the pool is cleared can be put (back) into the pool after
evictAll(). This means those connections will continue to
reference the old preferred / default network.

Bug: 17314604
Change-Id: Icbe20db77922cdbe1d6965785de7ee59495044de