Alter Android's connection pool behavior on network config changes
Rather than clearing out and closing the shared connection pool
each time the network changes with evictAll() this ensures that
later Http(s)URLConnections will reference a new ConnectionPool
instance.

This is an improvement over the current implementation:
1) It is obviously non-blocking.
2) It ensures that connections for the old network config
and are currently in use are not put back into the shared
pool.

Change-Id: I7b62f2bbd306c932dfd8d0ff4fff236ec04405bf
Bug: 17314604