Avoid a reverse DNS-lookup for a numeric proxy address
When the proxy is specified as an IP address the RouteSelector
would force a reverse-lookup. This can introduce extra latency
when the DNS is slow or missing. It also then turns the name
back into a set of IPs, which may result in the original IP
address not being given priority.

This is upstream commit:
https://github.com/square/okhttp/commit/9acd5e7b54fb55f62a5dd16865c30787dfba4c8b

Bug: 18327075
Change-Id: Idafd72f4d5b901e267081f924377db4ada9afe1b