Avoid a reverse DNS-lookup for a numeric proxy address
When the proxy is specified as an IP address the RouteSelector
would force a reverse-lookup. This can introduce extra latency
when the DNS is slow or missing. It also then turns the name
back into a set of IPs, which may result in the original IP
address not being given priority.

This is upstream commit:
https://github.com/square/okhttp/commit/9acd5e7b54fb55f62a5dd16865c30787dfba4c8b

Bug: 18567569
Bug: 18327075
(cherry picked from commit 797fdc6cc6bf16372e9464f189b535148f944ce9)

Change-Id: I52b4049095acafd2367530532d8cad7d75b6809a