Remove okhttp-static build target
There are no obvious dependencies on it.

Removing it should dispel any assumptions / implicit guarantees
that this fork of okhttp will work with any Android release. It enables
maintainers to relax / not code work-arounds when referencing methods
in libcore that may not be present on older Android releases.

For example, those methods marked with @since 1.7.

See also external/okhttpv2

Change-Id: I34eca60ad647b5328bc1c9323d350578bdd7dabe