Workaround for request header values with trailing '\r' or '\r\n'
Testing is showing an app that appears to
be setting a request header with a trailing '\r'. This change
strips the '\r' or '\r\n' iff it is at the end of the value.

(cherry picked from commit 8ef2daccd20f80b439c93af7d7985fb26d1f6b2f)

Bug: 26889631
Change-Id: I3a521a5272ad0c7cf3b018892516ced9e65f46ca